AKS: Disable actual invocation of kubeprod (temporarily)
kubeprod is failing using the new AKS code, when run as the jenkins
service principal (works when run locally).

This change disables running kubeprod on AKS so we can work from the
newer codebase.  This effectively makes the integration test useless
on AKS, so fixing/restoring this is a priority.