Simplify platform expectations and AKS interaction
Previously, the platform-specific code was expected to be idempotent.
This is difficult in Azure (and presumably other) platforms, since it
requires much greater permissions to search for an existing object
compared to just creating one.

The new code is dumber, only creates objects, and writes the result to
a local json file.  On re-runs (upgrades, etc) it will read the
existing json and skip any item that already exists.

While being simpler, this also allows power users to perform
equivalent setup in arbitrary ways and inject the desired values into
this json file before invoking `kubeprod`.