Functionality change for kubeprod install and init.
This change integrates the functionality of "kubeprod install" and
"kubeprod init" into a single "kubeprod install" command for simpler
usability.

For AKS, "kubeprod install" will create the required objects in
Azure (as part of the PreUpdate stage), then will create a jsonnet
manifests (root manifest) plus a JSON configuration file for the
AKS cluster being installed (as part of the Generate stage), and then
will run "kubecfg update" to deploy all the Kubernetes objects (as
part of the Update stage).

This effectively allows decoupling the AKS-specific part from the
k8s/kubecfg part allowwing for greater flexibility and easier to
customize cluster configuration.