alertmanager: remove templates volume
Creating a separate volume for alertmanager templates requires that we
need add another `configmap-reloader` to watch changes in the `templates`
volume and trigger a alertmanager reload. Having another reloader seems
to be overkill and therefore we decided that users should place their
custom alertmanager templates in the `config` volume instead.