Include client-go auth plugins
Include client-go auth plugins, required for (eg) gcp.

This was unusually difficult since the azure auth plugin conflicted
with the major version of the azure library we needed for our own
kubeprod aks code - triggering an update to most of our major
dependencies.  A more minimal change might have been possible, but the
final iteration chosen was:

```
govendor fetch \
  github.com/Azure/azure-sdk-for-go/...@v20 \
  github.com/Azure/go-autorest/...@v10 \
  github.com/ksonnet/kubecfg/... \
  k8s.io/api/...@kubernetes-1.10 \
  k8s.io/apimachinery/...@kubernetes-1.10 \
  k8s.io/client-go/...@kubernetes-1.10 \
  k8s.io/client-go/plugin/pkg/client/auth/azure@c49eef9ec85d226da87b0284e1f358a6d341110e
```

Note in particular that this includes an update to `kubecfg` master,
and from `client-go` v6 to v7 (aka kubernetes-1.10 branch), with a
cherrypicked fix to just the client-go azure auth plugin.