Squash to a single jsonnet tip per platform and remove `--platform`
The new update policy is "support at least two k8s versions with a
single version of BKPR".  This means we should remove the per
k8s-version tip files and squash to only a single tip per platform.

Combined with the recent change to per-platform subcommands, there is
no longer a need for the `--platform` flag, so remove that too.

\o/