Remove unused `nginx-internal` RBAC rule (#161)
We had an unused RBAC rule referring to an "nginx-internal" resource
that we don't use in BKPR manifests.  This PR cleans that up.

Also: this PR refactors the RBAC rule to point to the explicit
controller `--ingress-class` flag.  This makes it easier for users to
see where to change this value, and automatically changes the RBAC
rule to match.