Publish BKPR version in kubeprod/release configmap in-cluster
Create a "well known" ConfigMap in-cluster that serves as an answer
for the question "What version of BKPR is installed in this cluster?"

This ConfigMap is `release` in namespace `kubeprod`.  It is readable
by all authenticated users (`system:authenticated`).

This change also extends `kubeprod version` to fetch and report this
version.