ci: specify `DOCKER_CONFIG` env variable for kaniko
kaniko does not appear to pick docker hub credentials placed at
`/root/.docker/` anymore. As a result the release stage was not able to
push images to the docker registry. In this PR we explicitly specify the
path to the docker `config.json` file in the `DOCKER_CONFIG` variable.

We've also updated the kaniko version in this PR.

(cherry picked from commit 65a0b025af6381e32e2cface70e98afd97a71334)