tests: optimize retry timeouts
Successful passes of the ci tests generally conclude in 5 - 10mins. It has been observered that
when it takes longer the tests are bound to fail. Some of our tests use 30m retry timeouts
which is lot of time wasted hoping that a failing test will pass.

This commit optimizes the retry timeouts for all tests to more realistic values thereby reducing
the time spent by a job in the CI environment