ci: use github-integration plugin for PR's only (#587)
The github-integration plugin provides nice features that allow us to control the
behaviour of CI using github comments and labels. However we have been noticing issues
where build on branches and tags are randomly triggered.

To work around this issue, we are switching to processing only tags using this plugin while
processing branches and tags with the regular jenkins github integration.