Mark TLS tests as Pending
Unfortunately these tests are failing frequently from what appears to be
an inconsistent ACME challenge success rate. Upon further inspection,
we're seeing the challenges fail due to DNS resolution from the
LetsEncrypt side. Here's some output coming from running the tests in a
loop:

```
Status:
  Presented:   false
  Processing:  false
  Reason:      Error accepting authorization: acme: authorization error for test-ing-h8sk6.smuc.ci.bkpr.run: 400 urn:ietf:params:acme:error:dns: During secondary validation: No valid IP addresses found for test-ing-h8sk6.smuc.ci.bkpr.run
  State:       invalid
Events:
  Type     Reason     Age    From          Message
  ----     ------     ----   ----          -------
  Normal   Started    7m36s  cert-manager  Challenge scheduled for processing
  Normal   Presented  7m36s  cert-manager  Presented challenge using http-01 challenge mechanism
  Warning  Failed     4m48s  cert-manager  Accepting challenge authorization failed: acme: authorization error for test-ing-h8sk6.smuc.ci.bkpr.run: 400 urn:ietf:params:acme:error:dns: During secondary validation: No valid IP addresses found for test-ing-h8sk6.smuc.ci.bkpr.run
```

This failure seems to happen most often in GKE which implies the usage
of Google Cloud DNS. This doesn't happen as often in AKS or EKS (gut
estimate is that it happens 90% of the time in GKE). The above error
doesn't happen all the time though. Using a `dns-01` challenge has not
been attempted, but not sure if that would make this a more dependable
test.

Given that so many other things are out of date, we'll leave these tests
marked as Pending until after a modernisation effort is performed on
this repository.